+++ lib/eve_online.rb
@@ -64,6 +64,7 @@ require 'eve_online/esi/ancestries'
 require 'eve_online/esi/universe_types'
 require 'eve_online/esi/universe_type'
 require 'eve_online/esi/universe_regions'
+require 'eve_online/esi/universe_region'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'
@@ -107,6 +108,7 @@ require 'eve_online/esi/models/race'
 require 'eve_online/esi/models/bloodline'
 require 'eve_online/esi/models/faction'
 require 'eve_online/esi/models/ancestry'
+require 'eve_online/esi/models/region'
 
 module EveOnline
 end