+++ spec/eve_online/esi/universe_region_spec.rb
@@ -1,0 +1,106 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe EveOnline::ESI::UniverseRegion do
+  let(:options) { { id: 10_000_001 } }
+
+  subject { described_class.new(options) }
+
+  specify { expect(subject).to be_a(EveOnline::ESI::Base) }
+
+  specify { expect(described_class::API_ENDPOINT).to eq('https://esi.evetech.net/v1/universe/regions/%<region_id>s/?datasource=%<datasource>s&language=en-us') }
+
+  describe '#initialize' do
+    its(:token) { should eq(nil) }
+
+    its(:parser) { should eq(JSON) }
+
+    its(:_read_timeout) { should eq(60) }
+
+    its(:_open_timeout) { should eq(60) }
+
+    its(:datasource) { should eq('tranquility') }
+
+    its(:id) { should eq(10_000_001) }
+  end
+
+  describe '#model' do
+    let(:response) { double }
+
+    before { expect(subject).to receive(:response).and_return(response) }
+
+    let(:model) { double }
+
+    before do
+      #
+      # EveOnline::ESI::Models::Region.new(response) # => model
+      #
+      expect(EveOnline::ESI::Models::Region).to receive(:new).with(response).and_return(model)
+    end
+
+    specify { expect { subject.model }.not_to raise_error }
+
+    specify { expect { subject.model }.to change { subject.instance_variable_defined?(:@_memoized_model) }.from(false).to(true) }
+  end
+
+  describe '#as_json' do
+    let(:model) { double }
+
+    before { subject.instance_variable_set(:@_memoized_model, model) }
+
+    before { expect(model).to receive(:as_json) }
+
+    specify { expect { subject.as_json }.not_to raise_error }
+  end
+
+  describe '#constellations' do
+    let(:model) { double }
+
+    before { subject.instance_variable_set(:@_memoized_model, model) }
+
+    before { expect(model).to receive(:constellations) }
+
+    specify { expect { subject.constellations }.not_to raise_error }
+  end
+
+  describe '#description' do
+    let(:model) { double }
+
+    before { subject.instance_variable_set(:@_memoized_model, model) }
+
+    before { expect(model).to receive(:description) }
+
+    specify { expect { subject.description }.not_to raise_error }
+  end
+
+  describe '#name' do
+    let(:model) { double }
+
+    before { subject.instance_variable_set(:@_memoized_model, model) }
+
+    before { expect(model).to receive(:name) }
+
+    specify { expect { subject.name }.not_to raise_error }
+  end
+
+  describe '#region_id' do
+    let(:model) { double }
+
+    before { subject.instance_variable_set(:@_memoized_model, model) }
+
+    before { expect(model).to receive(:region_id) }
+
+    specify { expect { subject.region_id }.not_to raise_error }
+  end
+
+  describe '#scope' do
+    specify { expect(subject.scope).to eq(nil) }
+  end
+
+  describe '#url' do
+    specify do
+      expect(subject.url).to eq('https://esi.evetech.net/v1/universe/regions/10000001/?datasource=tranquility&language=en-us')
+    end
+  end
+end